renamed transformSettingsService files
